postings: fix term dict decoding for lengths >= 128
Two latent bugs, both only reachable with escape-coded lengths or block
starts of 128+ bytes (silent term corruption / OOB reads before this):
- readNextTermInBlock assigned the escape prefix/suffix length bytes from
  the signed char readByte(), sign-extending 128+ to negative.
- flushTerms wrote block-start terms with writeStr (vint length) while
  readers decode them with readPackedTerm (one size byte); the encodings
  diverge at 128+. Write with writePackedTerm to match.
TermsDictTest covers long shared prefixes, long suffixes, and 128+ byte
block-start terms via nextTerm/seek/seekCeil.

+++ src/solux/reader/TermsEnum.h
@@ -146,10 +146,10 @@ protected:
     auto prefixLen = code >> 5;
     auto suffixLen = (code & 0x1f) + 1;
     if (prefixLen == 7) {
-      prefixLen = termsIS.readByte();
+      prefixLen = (uint8_t)termsIS.readByte();
     }
     if (suffixLen == 32) {
-      suffixLen = termsIS.readByte();
+      suffixLen = (uint8_t)termsIS.readByte();
     }
 
     auto [data, len] = currTerm.unpack();
